requirements.txt
# Multimodal GenAI Studio - Dependencies

# Core Framework
gradio>=4.0.0
python-dotenv>=1.0.0

# AI/ML Libraries
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.7.0

# Image Generation
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.1.0

# Audio Processing
openai-whisper>=20231117
pydub>=0.25.1
gTTS>=2.4.0
soundfile>=0.12.1
librosa>=0.10.1

# Multimodal
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-openai>=0.0.2

# Utilities
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.12.0

